[Refactor] FileComponent 리팩토링 #28
