+++ src/main/java/me/snaptime/snap/component/FileComponent.java
@@ -1,13 +1,10 @@
-package me.snaptime.snap.service;
+package me.snaptime.snap.component;
 
 import me.snaptime.snap.data.dto.file.WritePhotoToFileSystemResult;
 
-import javax.crypto.SecretKey;
-
-public interface PhotoService {
-    byte[] downloadPhotoFromFileSystem(String filePath, SecretKey secretKey);
+public interface FileComponent {
     byte[] downloadPhotoFromFileSystem(String filePath);
-    void deletePhoto(Long id);
+    void deletePhoto(String fileName);
     byte[] getPhotoByte(String filePath);
     void updateFileSystemPhoto(String filePath, byte[] fileBytes);
     WritePhotoToFileSystemResult writePhotoToFileSystem(String fileName, String contentType, byte[] fileBytes);
