+++ src/main/java/me/snaptime/snap/controller/PhotoController.java
@@ -4,13 +4,14 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.Parameters;
 import io.swagger.v3.oas.annotations.tags.Tag;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
-import me.snaptime.common.jwt.JwtProvider;
 import me.snaptime.snap.service.SnapService;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -19,7 +20,6 @@ import org.springframework.web.bind.annotation.*;
 @Tag(name = "[Photo] Photo API")
 public class PhotoController {
     private final SnapService snapService;
-    private final JwtProvider jwtProvider;
 
     @Operation(summary = "Photo 조회", description = "")
     @Parameters({
@@ -28,11 +28,11 @@ public class PhotoController {
     @GetMapping
     public ResponseEntity<byte[]> findPhoto(
             final @RequestParam("fileName") String fileName,
-            final @RequestParam("isEncrypted") boolean isEncrypted,
-            final HttpServletRequest request
+            final @RequestParam("isEncrypted") boolean isEncrypted
     ) {
-        String token = jwtProvider.resolveToken(request);
-        String uId = jwtProvider.getUsername(token);
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        UserDetails userDetails = (UserDetails) authentication.getPrincipal();
+        String uId = userDetails.getUsername();
         return ResponseEntity.status(HttpStatus.OK).contentType(MediaType.IMAGE_PNG).body(
                 snapService.downloadPhotoFromFileSystem(fileName, uId, isEncrypted)
         );
